backend/requirement.txt
fastapi==0.60.*
python-multipart==0.0.5
tensorflow==2.2.*
uvicorn==0.11.8
Pillow==7.2.*